Proyecto P2P/src/requirements.txt
fastapi
uvicorn[standard]
grpcio
grpcio-tools
protobuf
aiofiles
requests
python-dotenv
